requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: linux-64
# created-by: conda 24.11.1
_libgcc_mutex=0.1=main
_openmp_mutex=5.1=1_gnu
accelerate=1.4.0=pypi_0
aiohappyeyeballs=2.4.6=pypi_0
aiohttp=3.11.13=pypi_0
aiosignal=1.3.2=pypi_0
annotated-types=0.7.0=pypi_0
anyio=4.8.0=pypi_0
asgiref=3.8.1=pypi_0
asttokens=3.0.0=pyhd8ed1ab_1
attrs=25.1.0=pypi_0
beautifulsoup4=4.13.3=pypi_0
bitsandbytes=0.45.3=pypi_0
blinker=1.9.0=pypi_0
bzip2=1.0.8=h5eee18b_6
ca-certificates=2025.1.31=hbcca054_0
certifi=2025.1.31=pypi_0
charset-normalizer=3.4.1=pypi_0
click=8.1.8=pypi_0
comm=0.2.2=pyhd8ed1ab_1
contourpy=1.3.1=pypi_0
cut-cross-entropy=25.1.1=pypi_0
cycler=0.12.1=pypi_0
datasets=3.3.2=pypi_0
debugpy=1.8.11=py312h6a678d5_0
decorator=5.2.1=pyhd8ed1ab_0
diffusers=0.32.2=pypi_0
dill=0.3.8=pypi_0
docker-pycreds=0.4.0=pypi_0
docstring-parser=0.16=pypi_0
duckduckgo-search=7.5.0=pypi_0
exceptiongroup=1.2.2=pyhd8ed1ab_1
executing=2.1.0=pyhd8ed1ab_1
expat=2.6.4=h6a678d5_0
filelock=3.17.0=pypi_0
flask=3.1.0=pypi_0
flask-cors=5.0.1=pypi_0
fonttools=4.56.0=pypi_0
frozenlist=1.5.0=pypi_0
fsspec=2024.12.0=pypi_0
gitdb=4.0.12=pypi_0
gitpython=3.1.44=pypi_0
h11=0.14.0=pypi_0
hf-transfer=0.1.9=pypi_0
huggingface-hub=0.29.1=pypi_0
idna=3.10=pypi_0
importlib-metadata=8.6.1=pyha770c72_0
ipykernel=6.29.5=pyh3099207_0
ipython=8.32.0=pyh907856f_0
itsdangerous=2.2.0=pypi_0
jedi=0.19.2=pyhd8ed1ab_1
jinja2=3.1.5=pypi_0
jupyter_client=8.6.3=pyhd8ed1ab_1
jupyter_core=5.7.2=pyh31011fe_1
kiwisolver=1.4.8=pypi_0
krb5=1.21.3=h143b758_0
ld_impl_linux-64=2.40=h12ee557_0
libedit=3.1.20230828=h5eee18b_0
libffi=3.4.4=h6a678d5_1
libgcc=14.2.0=h767d61c_2
libgcc-ng=14.2.0=h69a702a_2
libgomp=14.2.0=h767d61c_2
libsodium=1.0.20=h4ab18f5_0
libstdcxx=14.2.0=h8f9b012_2
libstdcxx-ng=11.2.0=h1234567_1
libuuid=1.41.5=h5eee18b_0
lxml=5.3.1=pypi_0
markdown-it-py=3.0.0=pypi_0
markdownify=1.0.0=pypi_0
markupsafe=3.0.2=pypi_0
matplotlib=3.10.0=pypi_0
matplotlib-inline=0.1.7=pyhd8ed1ab_1
mdurl=0.1.2=pypi_0
mpmath=1.3.0=pypi_0
multidict=6.1.0=pypi_0
multiprocess=0.70.16=pypi_0
ncurses=6.4=h6a678d5_0
nest-asyncio=1.6.0=pyhd8ed1ab_1
networkx=3.4.2=pypi_0
numpy=2.2.3=pypi_0
nvidia-cublas-cu12=12.4.5.8=pypi_0
nvidia-cuda-cupti-cu12=12.4.127=pypi_0
nvidia-cuda-nvrtc-cu12=12.4.127=pypi_0
nvidia-cuda-runtime-cu12=12.4.127=pypi_0
nvidia-cudnn-cu12=9.1.0.70=pypi_0
nvidia-cufft-cu12=11.2.1.3=pypi_0
nvidia-curand-cu12=10.3.5.147=pypi_0
nvidia-cusolver-cu12=11.6.1.9=pypi_0
nvidia-cusparse-cu12=12.3.1.170=pypi_0
nvidia-cusparselt-cu12=0.6.2=pypi_0
nvidia-nccl-cu12=2.21.5=pypi_0
nvidia-nvjitlink-cu12=12.4.127=pypi_0
nvidia-nvtx-cu12=12.4.127=pypi_0
openssl=3.4.1=h7b32b05_0
packaging=24.2=pyhd8ed1ab_2
pandas=2.2.3=pypi_0
parso=0.8.4=pyhd8ed1ab_1
peft=0.14.0=pypi_0
pexpect=4.9.0=pyhd8ed1ab_1
pickleshare=0.7.5=pyhd8ed1ab_1004
pillow=11.1.0=pypi_0
pip=25.0=py312h06a4308_0
platformdirs=4.3.6=pyhd8ed1ab_1
primp=0.14.0=pypi_0
prompt-toolkit=3.0.50=pyha770c72_0
propcache=0.3.0=pypi_0
protobuf=3.20.3=pypi_0
psutil=7.0.0=pypi_0
ptyprocess=0.7.0=pyhd8ed1ab_1
pure_eval=0.2.3=pyhd8ed1ab_1
pyarrow=19.0.1=pypi_0
pydantic=2.10.6=pypi_0
pydantic-core=2.27.2=pypi_0
pygments=2.19.1=pyhd8ed1ab_0
pyparsing=3.2.1=pypi_0
python=3.12.9=h5148396_0
python-dateutil=2.9.0.post0=pyhff2d567_1
python-dotenv=1.0.1=pypi_0
pytz=2025.1=pypi_0
pyyaml=6.0.2=pypi_0
pyzmq=26.2.0=py312h6a678d5_0
readline=8.2=h5eee18b_0
regex=2024.11.6=pypi_0
requests=2.32.3=pypi_0
rich=13.9.4=pypi_0
safetensors=0.5.2=pypi_0
sentencepiece=0.2.0=pypi_0
sentry-sdk=2.22.0=pypi_0
setproctitle=1.3.5=pypi_0
setuptools=75.8.0=py312h06a4308_0
shtab=1.7.1=pypi_0
six=1.17.0=pyhd8ed1ab_0
smmap=5.0.2=pypi_0
smolagents=1.9.2=pypi_0
sniffio=1.3.1=pypi_0
soupsieve=2.6=pypi_0
sqlite=3.45.3=h5eee18b_0
stack_data=0.6.3=pyhd8ed1ab_1
sympy=1.13.1=pypi_0
tiktoken=0.9.0=pypi_0
tk=8.6.14=h39e8969_0
tokenizers=0.21.0=pypi_0
torch=2.6.0=pypi_0
torchvision=0.21.0=pypi_0
tornado=6.4.2=py312h5eee18b_0
tqdm=4.67.1=pypi_0
traitlets=5.14.3=pyhd8ed1ab_1
transformers=4.49.0=pypi_0
triton=3.2.0=pypi_0
trl=0.15.2=pypi_0
typeguard=4.4.2=pypi_0
typing_extensions=4.12.2=pyha770c72_1
tyro=0.9.16=pypi_0
tzdata=2025.1=pypi_0
unsloth=2025.2.15=pypi_0
unsloth-zoo=2025.2.7=pypi_0
urllib3=2.3.0=pypi_0
uvicorn=0.34.0=pypi_0
wandb=0.19.7=pypi_0
watchfiles=1.0.4=pypi_0
wcwidth=0.2.13=pyhd8ed1ab_1
werkzeug=3.1.3=pypi_0
wheel=0.45.1=py312h06a4308_0
xformers=0.0.29.post3=pypi_0
xxhash=3.5.0=pypi_0
xz=5.6.4=h5eee18b_1
yarl=1.18.3=pypi_0
zeromq=4.3.5=h3b0a872_7
zipp=3.21.0=pyhd8ed1ab_1
zlib=1.2.13=h5eee18b_1
